Изменён тип OptionBolts и обновлён парсинг столбца
В классе Row свойство OptionBolts изменено с int на double для поддержки дробных значений. В BuildBoltsColumnMap ключ для OptionBolts изменён с "version" на "Option". Парсинг значения по-прежнему через ParseInt, но теперь результат сохраняется в double.

+++ ConvertData/Domain/Row.cs
@@ -271,7 +271,7 @@ internal sealed class Row
     /// <summary>
     /// "Bolts": Версия использования болтов
     /// </summary>
-    public int OptionBolts { get; set; } = 0;
+    public double OptionBolts { get; set; } = 0;
     #endregion
